[FAB-6886] SDK logger new utility function
Current utility 'IsEnabledFor' needs module name
as an argument to find if given level is enabled
for given module, which forces clients to preserve
logger module names for this use.

This new utility, IsEnabledForLogger - takes logger itself
as an argument and it will find module name by accessing
logger argument.


Change-Id: I0242667c12428a7a99e85184f4d7e639e3beb84e
Signed-off-by: Sudesh Shetty <sudesh.shetty@securekey.com>