FAB-8101 - .json file extension with filekeyvaluestore
FAB-7513 (change https://gerrit.hyperledger.org/r/#/c/17547/ )
changed the default file extension of the new files produced by
the filekeyvaluestore. Previously it was ".json", and now there
is no extension (the key becomes the filename). This ticket
demonstrates how to construct a new filekeyvaluestore that adds
".json" to the store file names, to keep compatibility with
previously created stores.

Change-Id: Iddd20a78f941d2d8449e3ef8a970c6d89ba8c8d2
Signed-off-by: Aleksandar Likic <aleksandar.likic@securekey.com>