[FAB-8546] Fixed race condition in event client
A race condition existed during the shutdown of
the event dispatcher such that an event was being
posted over the event channel while it was being
closed. The fix is to not close the dispatcher event
channel but to check the state of the dispatcher
after handling each event and exit the loop if the
state is set to stopped.

Change-Id: I5e6542f3118458a3cb497d6969ef030d4c9c5899
Signed-off-by: Bob Stasyszyn <Bob.Stasyszyn@securekey.com>