[FAB-8954] Fix index out of range panic
The index out of range occurs because the
peer retriever is called multiple times
during Resolve and sometomes it returns a
different sized set of peers (this can
happen if a peer goes down or comes back
up during the operation). This patch fixes
the problem by feeding the peer group
resolver a peer retriever that always
returns the same set of peers for the
duration of the call to Resolve.

Change-Id: I265943f6c2a5339ce7a411620aa016c6d5078a9c
Signed-off-by: Bob Stasyszyn <Bob.Stasyszyn@securekey.com>