[FABG-761] Pluggable peer resolvers for event client
The event client now allows you to specify which peer
resolver to use when selecting peers to connect to. To
peer resolver also decides when to disconnect to from
the connected peer.

Also added a vanilla peer resolver that simply load
balances between peers when choosing the initial peer
to connect to.

Change-Id: Ia0ef4db4faa2c2ea9c2535ac8ccaee7cfbe6c238
Signed-off-by: Bob Stasyszyn <Bob.Stasyszyn@securekey.com>