FABG-888 - Add ID property to CA definition
This change adds an ID to CA configuration definition,
to be used as CA identity. Before this change, SDK was
using CA name as CA identity.

This change is mostly internal to SDK. SDK yaml
configuration file is not affected. During
SDK coniguration initialization, CA ID is internally
set to the value of the CA configuration key.

Change-Id: I98e9b34d21b3a28cf7b6d16d945537f61985a151
Signed-off-by: Aleksandar Likic <aleksandar.likic@securekey.com>