FABG-936 Support minimal ccp for fabric samples (#74)
The Fabcar sample in fabric-samples contains a minimal CCP definition with just a single gateway peer entry and no channel or orderer information.  It is also written to be run on a single dev system whereby hostnames need to be translated to localhost.  This commit adds support to the gateway package for:
- automatic creation of channel definition on invocation of `getNetwork(channelName)`
- support for translation of discovered peers/orderers to localhost via environment variable (DISCOVERY_AS_LOCALHOST)
Both of these align the behaviour with the Node and Java SDKs and are required to implement the Fabcar sample.

Signed-off-by: andrew-coleman <andrew_coleman@uk.ibm.com>