Fix the Ability to Override CredentialStorePath
The existing implementation pulls the value from the config
file. If a user overrides the CredentialStorePath it does not
supersede the value defined in client.config.credentialStore.Path.
Moreover, if a user performs a configless configuration of the SDK
where there is no configfile, the value is read in as "", and an
inMemoryStore is used as opposed to the expected override.

This change ensures the overridden value supersedes the config value
but retains the config value if no override is passed.

Signed-off-by: Brett Logan <brett.t.logan@ibm.com>