FABG-1018 TRANSIENT_FAILURE in SubmitTransaction
When calling SubmitTransaction in the gateway package immediately following GetNetwork, it fails with TRANSIENT_FAILURE
The orderer log indicates bad TLS certificate.

This is a timing issue whereby the discovery information for the channel config was not available early enough.
Code had been added to force this to be read during channel initialization.

Signed-off-by: andrew-coleman <andrew_coleman@uk.ibm.com>