Get proposal response payload from chaincode action (#206)
This is the correct chaincode transaction response payload location. The response payload field is intended to contain metadata but the peer currently copies the transaction response payload here for convenience. This causes failures due to gRPC message size limits for large response payloads and so may be removed in the future.

Also changed a few build scripts to honour the GO_CMD environment variable.

Signed-off-by: Mark S. Lewis <mark_lewis@uk.ibm.com>